requirements.txt
# web server
uvicorn

# web framework
fastapi

# models declarations
pydantic

# for jwt tokens
python-jose[cryptography]

# for password hashing
passlib[bcrypt]

# mongo adapter
pymongo

# for multipart form-data
python-multipart

# File support for asyncio
aiofiles

# for image processing
opencv-python

# matrix operations (for opencv)
numpy

# utils for work with images
imutils

# load variables from env file
python-dotenv
